Allowing user to select coffee name rather than id
A coffee name will make more sense to a user rather than an ID.
Therefore the coffees model must be called in the get controller to pass
all coffee data into the edit view. This is then put into a drop down of
options where the user can select a coffee name. The coffee ID is then
posted back to the update then the database is updated.