Retrieving and checking data from sign up form
This javascript file prevents the data from submitting and performs
checks on the data to see if there is any issues in the form. This file
retrieves the data using document query selector. It then will post the
data to the api/users route where the data is all individually checked
and a JSON string is passed back to this javascript file.