Validating user in put on login form
Once this file retrieves the posted data from the javascript file it
checks firstly that it recieves data. If data is there it will then
check if user already exists and will output an error message in console
if there is. If not, it will check the password and confirm passwords
match then will ensure the passwords meet the regex criteria set. The
data is then returned the javascript file in json format where it will
display success if all requirements are met and if not, the error
message will be displayed.
There is also a check at the moment to ensure if the user selects that
they want to be an admin, their password must have a certain format.