Adding session check to limit what user can see
If user is not logged, they will only be able to see the price. If they
are logged in they can see the add to cart button and add a quantity. If
the admin is logged in, they will see add to cart, the quantity form
input, the delte button and edit button.